Sanitize image orientation and texture origin handling
Both STB and ANARI expect the images to be top down, no need to flip on load with STB.
ANARI and ASSIMP actually disagree on UV origin, so make sure ASSIMP flip UVs on load.
Same with MikkT, which expects the texture coordinates to be starting lower-left.

This is a no-op as far as the assimp importer is concerned, but this makes sure the
behavior is intentional and other tools, such as VisRTX tangent computation,
can be used reliably, with all importers.